+++ .gitignore
@@ -1,2 +1,3 @@
 # Project exclude paths
 /out/
+/src/Test.java
