+++ src/Main.java
@@ -1,17 +1,11 @@
+import java.util.Scanner;
+
 public class Main {
 
-    public static void main(String[] args) throws Exception {
-	    Book TJ = new Book("Tõde ja õigus IV", "Anton Hansen Tammsaare");
-
-	    TJ.setISBN("9789949664504");
-	    TJ.setStatus(Status.READING);
-        TJ.setGenre("Draama");
-        TJ.setPublicationDate("1964");
-        System.out.println(TJ);
-
-        Library minuRaamatukogu = new Library("Minu Raamatukogu");
-
-        minuRaamatukogu.addBook(TJ);
-        System.out.println(minuRaamatukogu);
+    public static void main(String[] args) {
+        Scanner sc = new Scanner(System.in);
+        System.out.println("sisesta uue raamatu pealkiri: ");
+        String pealkiri = sc.nextLine();
+        System.out.println(pealkiri);
     }
 }
